refactor(WalletForm): extract settings card
